+++ builder/utils/util_math.py
@@ -15,3 +15,8 @@ def int_ceil(a: int, b: int) -> int:
     '''
     return -(-assertion.is_int(a) // assertion.is_int(b))
 
+
+def safe_divided(a: (int, float), b: (int, float)) -> (int, float):
+    ''' Except zero divide
+    '''
+    return a / b if b else 0